Remove AntlrParserBase; make AntlrParserConverter inherited from AntlrParser
